+++ src/runtime-rs/crates/hypervisor/src/device/driver/virtio_blk_modern.rs
@@ -6,7 +6,6 @@
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
-use crate::device::driver::BlockDeviceAio;
 use crate::device::pci_path::PciPath;
 use crate::device::topology::PCIeTopology;
 use crate::device::util::do_decrease_count;
@@ -16,6 +15,63 @@ use crate::device::DeviceType;
 use crate::Hypervisor as hypervisor;
 use anyhow::{Context, Result};
 use async_trait::async_trait;
+
+/// VIRTIO_BLOCK_PCI indicates block driver is virtio-pci based
+pub const VIRTIO_BLOCK_PCI: &str = "virtio-blk-pci";
+pub const VIRTIO_BLOCK_MMIO: &str = "virtio-blk-mmio";
+pub const VIRTIO_BLOCK_CCW: &str = "virtio-blk-ccw";
+pub const VIRTIO_PMEM: &str = "virtio-pmem";
+
+#[derive(Clone, Copy, Debug, Default)]
+pub enum BlockDeviceAio {
+    // IoUring is the Linux io_uring I/O implementation.
+    #[default]
+    IoUring,
+
+    // Native is the native Linux AIO implementation.
+    Native,
+
+    // Threads is the pthread asynchronous I/O implementation.
+    Threads,
+}
+
+impl BlockDeviceAio {
+    pub fn new(aio: &str) -> Self {
+        match aio {
+            "native" => BlockDeviceAio::Native,
+            "threads" => BlockDeviceAio::Threads,
+            _ => BlockDeviceAio::IoUring,
+        }
+    }
+}
+
+impl std::fmt::Display for BlockDeviceAio {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let to_string = match *self {
+            BlockDeviceAio::Native => "native".to_string(),
+            BlockDeviceAio::Threads => "threads".to_string(),
+            _ => "iouring".to_string(),
+        };
+        write!(f, "{to_string}")
+    }
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub enum BlockDeviceFormat {
+    #[default]
+    Raw,
+    Vmdk,
+}
+
+impl std::fmt::Display for BlockDeviceFormat {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let to_string = match *self {
+            BlockDeviceFormat::Raw => "raw".to_string(),
+            BlockDeviceFormat::Vmdk => "vmdk".to_string(),
+        };
+        write!(f, "{to_string}")
+    }
+}
 
 #[derive(Debug, Clone, Default)]
 pub struct BlockConfigModern {
@@ -31,6 +87,9 @@ pub struct BlockConfigModern {
 
     /// Don't close `path_on_host` file when dropping the device.
     pub no_drop: bool,
+
+    /// raw, vmdk, etc. And default to raw if not set.
+    pub format: BlockDeviceFormat,
 
     /// Specifies cache-related options for block devices.
     /// Denotes whether use of O_DIRECT (bypass the host page cache) is enabled.
